refactor(LottoMatch) : 예외값을 처리하는 NONE_MATCH 객체 생성
