+++ src/main/java/lotto/domain/LottoMatch.java
@@ -1,7 +1,8 @@
-package lottogame;
+package lotto.domain;
 
 import java.util.Collections;
 import java.util.Map;
+import java.util.Optional;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
@@ -11,7 +12,8 @@ public enum LottoMatch {
     FOUR_MATCH(new LottoMatchKey(false, 4), 50_000),
     FIVE_MATCH(new LottoMatchKey(false, 5), 1_500_000),
     FIVE_BONUS_MATCH(new LottoMatchKey(true, 5), 30_000_000),
-    SIX_MATCH(new LottoMatchKey(false, 6), 2_000_000_000);
+    SIX_MATCH(new LottoMatchKey(false, 6), 2_000_000_000),
+    NONE_MATCH(new LottoMatchKey(true, 0), 0);
 
     private final LottoMatchKey lottoMatchKey;
     private final int prize;
@@ -22,7 +24,7 @@ public enum LottoMatch {
     }
 
     public static LottoMatch find(LottoMatchKey lottoMatchKey) {
-        return MAP.get(lottoMatchKey);
+        return Optional.ofNullable(MAP.get(lottoMatchKey)).orElse(NONE_MATCH);
     }
 
     private static final Map<LottoMatchKey, LottoMatch> MAP =
