+++ src/main/java/lotto/domain/LottoMatchKey.java
@@ -1,4 +1,4 @@
-package lottogame;
+package lotto.domain;
 
 import java.util.Objects;
